Update MainMenu.cs
Made several changes to the class
Added LoadFarmScene function as listener of the onClick function of the start button, all in the script Start function.
Renamed ReadTextfiel to SaveFarmName
Changed accesibility level of LoadFarmSCene function from public for private.
Added a condition to check if text has been written in the input field, printing a warning if no text is found, otherwise the farm scene is loaded. (saving farms name not yet implemented)